Plotting geo_em from geogrid.exe to test the WPS configuration
Setting up the domains is a challenging task in WRF. This code allows to visualize the geo_em file generated from geogrid.exe to visualize with a shape file of your choice to see if the domain configuration is done as you desire. Plotgrid.ncl in wps/util does not provide the visualization of a shape file and therefore it is hard to setup the domains precisely.